+++ packages/contracts/contracts/SortedTroves.sol
@@ -11,49 +11,53 @@ import "./Dependencies/CheckContract.sol";
 import "./Dependencies/console.sol";
 
 /**
-* A sorted doubly linked list with nodes sorted in descending order.
-*
-* Nodes map to active Troves in the system - the ID property is the address of a Trove owner.
-* Nodes are ordered according to their current nominal individual collateral ratio (NICR),
-* which is like the ICR but without the price, i.e., just collateral / debt.
-*
-* The list optionally accepts insert position hints.
-*
-* NICRs are computed dynamically at runtime, and not stored on the Node. This is because NICRs of active Troves
-* change dynamically as liquidation events occur.
-*
-* The list relies on the fact that liquidation events preserve ordering: a liquidation decreases the NICRs of all active Troves,
-* but maintains their order. A node inserted based on current NICR will maintain the correct position,
-* relative to it's peers, as rewards accumulate, as long as it's raw collateral and debt have not changed.
-* Thus, Nodes remain sorted by current NICR.
-*
-* Nodes need only be re-inserted upon a Trove operation - when the owner adds or removes collateral or debt
-* to their position.
-*
-* The list is a modification of the following audited SortedDoublyLinkedList:
-* https://github.com/livepeer/protocol/blob/master/contracts/libraries/SortedDoublyLL.sol
-*
-*
-* Changes made in the Zero implementation:
-*
-* - Keys have been removed from nodes
-*
-* - Ordering checks for insertion are performed by comparing an NICR argument to the current NICR, calculated at runtime.
-*   The list relies on the property that ordering by ICR is maintained as the ETH:USD price varies.
-*
-* - Public functions with parameters have been made internal to save gas, and given an external wrapper function for external access
-*/
+ * A sorted doubly linked list with nodes sorted in descending order.
+ *
+ * Nodes map to active Troves in the system - the ID property is the address of a Trove owner.
+ * Nodes are ordered according to their current nominal individual collateral ratio (NICR),
+ * which is like the ICR but without the price, i.e., just collateral / debt.
+ *
+ * The list optionally accepts insert position hints.
+ *
+ * NICRs are computed dynamically at runtime, and not stored on the Node. This is because NICRs of active Troves
+ * change dynamically as liquidation events occur.
+ *
+ * The list relies on the fact that liquidation events preserve ordering: a liquidation decreases the NICRs of all active Troves,
+ * but maintains their order. A node inserted based on current NICR will maintain the correct position,
+ * relative to it's peers, as rewards accumulate, as long as it's raw collateral and debt have not changed.
+ * Thus, Nodes remain sorted by current NICR.
+ *
+ * Nodes need only be re-inserted upon a Trove operation - when the owner adds or removes collateral or debt
+ * to their position.
+ *
+ * The list is a modification of the following audited SortedDoublyLinkedList:
+ * https://github.com/livepeer/protocol/blob/master/contracts/libraries/SortedDoublyLL.sol
+ *
+ *
+ * Changes made in the Zero implementation:
+ *
+ * - Keys have been removed from nodes
+ *
+ * - Ordering checks for insertion are performed by comparing an NICR argument to the current NICR, calculated at runtime.
+ *   The list relies on the property that ordering by ICR is maintained as the ETH:USD price varies.
+ *
+ * - Public functions with parameters have been made internal to save gas, and given an external wrapper function for external access
+ */
 contract SortedTroves is SortedTrovesStorage, CheckContract, ISortedTroves {
     using SafeMath for uint256;
 
     event TroveManagerAddressChanged(address _troveManagerAddress);
     event BorrowerOperationsAddressChanged(address _borrowerOperationsAddress);
-    event NodeAdded(address _id, uint _NICR);
+    event NodeAdded(address _id, uint256 _NICR);
     event NodeRemoved(address _id);
 
     // --- Dependency setters ---
 
-    function setParams(uint256 _size, address _troveManagerAddress, address _borrowerOperationsAddress) external override onlyOwner {
+    function setParams(
+        uint256 _size,
+        address _troveManagerAddress,
+        address _borrowerOperationsAddress
+    ) external override onlyOwner {
         require(_size > 0, "SortedTroves: Size can’t be zero");
         checkContract(_troveManagerAddress);
         checkContract(_borrowerOperationsAddress);
@@ -65,8 +69,6 @@ contract SortedTroves is SortedTrovesStorage, CheckContract, ISortedTroves {
 
         emit TroveManagerAddressChanged(_troveManagerAddress);
         emit BorrowerOperationsAddressChanged(_borrowerOperationsAddress);
-
-        
     }
 
     /**
@@ -77,14 +79,25 @@ contract SortedTroves is SortedTrovesStorage, CheckContract, ISortedTroves {
      * @param _nextId Id of next node for the insert position
      */
 
-    function insert (address _id, uint256 _NICR, address _prevId, address _nextId) external override {
+    function insert(
+        address _id,
+        uint256 _NICR,
+        address _prevId,
+        address _nextId
+    ) external override {
         ITroveManager troveManagerCached = troveManager;
 
         _requireCallerIsBOorTroveM(troveManagerCached);
         _insert(troveManagerCached, _id, _NICR, _prevId, _nextId);
     }
 
-    function _insert(ITroveManager _troveManager, address _id, uint256 _NICR, address _prevId, address _nextId) internal {
+    function _insert(
+        ITroveManager _troveManager,
+        address _id,
+        uint256 _NICR,
+        address _prevId,
+        address _nextId
+    ) internal {
         // List must not be full
         require(!isFull(), "SortedTroves: List is full");
         // List must not already contain node
@@ -103,7 +116,7 @@ contract SortedTroves is SortedTrovesStorage, CheckContract, ISortedTroves {
             (prevId, nextId) = _findInsertPosition(_troveManager, _NICR, prevId, nextId);
         }
 
-         data.nodes[_id].exists = true;
+        data.nodes[_id].exists = true;
 
         if (prevId == address(0) && nextId == address(0)) {
             // Insert as head and tail
@@ -184,7 +197,12 @@ contract SortedTroves is SortedTrovesStorage, CheckContract, ISortedTroves {
      * @param _prevId Id of previous node for the new insert position
      * @param _nextId Id of next node for the new insert position
      */
-    function reInsert(address _id, uint256 _newNICR, address _prevId, address _nextId) external override {
+    function reInsert(
+        address _id,
+        uint256 _newNICR,
+        address _prevId,
+        address _nextId
+    ) external override {
         ITroveManager troveManagerCached = troveManager;
 
         _requireCallerIsBOorTroveM(troveManagerCached);
@@ -270,11 +288,20 @@ contract SortedTroves is SortedTrovesStorage, CheckContract, ISortedTroves {
      * @param _prevId Id of previous node for the insert position
      * @param _nextId Id of next node for the insert position
      */
-    function validInsertPosition(uint256 _NICR, address _prevId, address _nextId) external view override returns (bool) {
+    function validInsertPosition(
+        uint256 _NICR,
+        address _prevId,
+        address _nextId
+    ) external view override returns (bool) {
         return _validInsertPosition(troveManager, _NICR, _prevId, _nextId);
     }
 
-    function _validInsertPosition(ITroveManager _troveManager, uint256 _NICR, address _prevId, address _nextId) internal view returns (bool) {
+    function _validInsertPosition(
+        ITroveManager _troveManager,
+        uint256 _NICR,
+        address _prevId,
+        address _nextId
+    ) internal view returns (bool) {
         if (_prevId == address(0) && _nextId == address(0)) {
             // `(null, null)` is a valid insert position if the list is empty
             return isEmpty();
@@ -286,9 +313,10 @@ contract SortedTroves is SortedTrovesStorage, CheckContract, ISortedTroves {
             return data.tail == _prevId && _NICR <= _troveManager.getNominalICR(_prevId);
         } else {
             // `(_prevId, _nextId)` is a valid insert position if they are adjacent nodes and `_NICR` falls between the two nodes' NICRs
-            return data.nodes[_prevId].nextId == _nextId &&
-                   _troveManager.getNominalICR(_prevId) >= _NICR &&
-                   _NICR >= _troveManager.getNominalICR(_nextId);
+            return
+                data.nodes[_prevId].nextId == _nextId &&
+                _troveManager.getNominalICR(_prevId) >= _NICR &&
+                _NICR >= _troveManager.getNominalICR(_nextId);
         }
     }
 
@@ -298,7 +326,11 @@ contract SortedTroves is SortedTrovesStorage, CheckContract, ISortedTroves {
      * @param _NICR Node's NICR
      * @param _startId Id of node to start descending the list from
      */
-    function _descendList(ITroveManager _troveManager, uint256 _NICR, address _startId) internal view returns (address, address) {
+    function _descendList(
+        ITroveManager _troveManager,
+        uint256 _NICR,
+        address _startId
+    ) internal view returns (address, address) {
         // If `_startId` is the head, check if the insert position is before the head
         if (data.head == _startId && _NICR >= _troveManager.getNominalICR(_startId)) {
             return (address(0), _startId);
@@ -308,7 +340,9 @@ contract SortedTroves is SortedTrovesStorage, CheckContract, ISortedTroves {
         address nextId = data.nodes[prevId].nextId;
 
         // Descend the list until we reach the end or until we find a valid insert position
-        while (prevId != address(0) && !_validInsertPosition(_troveManager, _NICR, prevId, nextId)) {
+        while (
+            prevId != address(0) && !_validInsertPosition(_troveManager, _NICR, prevId, nextId)
+        ) {
             prevId = data.nodes[prevId].nextId;
             nextId = data.nodes[prevId].nextId;
         }
@@ -322,7 +356,11 @@ contract SortedTroves is SortedTrovesStorage, CheckContract, ISortedTroves {
      * @param _NICR Node's NICR
      * @param _startId Id of node to start ascending the list from
      */
-    function _ascendList(ITroveManager _troveManager, uint256 _NICR, address _startId) internal view returns (address, address) {
+    function _ascendList(
+        ITroveManager _troveManager,
+        uint256 _NICR,
+        address _startId
+    ) internal view returns (address, address) {
         // If `_startId` is the tail, check if the insert position is after the tail
         if (data.tail == _startId && _NICR <= _troveManager.getNominalICR(_startId)) {
             return (_startId, address(0));
@@ -332,7 +370,9 @@ contract SortedTroves is SortedTrovesStorage, CheckContract, ISortedTroves {
         address prevId = data.nodes[nextId].prevId;
 
         // Ascend the list until we reach the end or until we find a valid insertion point
-        while (nextId != address(0) && !_validInsertPosition(_troveManager, _NICR, prevId, nextId)) {
+        while (
+            nextId != address(0) && !_validInsertPosition(_troveManager, _NICR, prevId, nextId)
+        ) {
             nextId = data.nodes[nextId].prevId;
             prevId = data.nodes[nextId].prevId;
         }
@@ -346,11 +386,20 @@ contract SortedTroves is SortedTrovesStorage, CheckContract, ISortedTroves {
      * @param _prevId Id of previous node for the insert position
      * @param _nextId Id of next node for the insert position
      */
-    function findInsertPosition(uint256 _NICR, address _prevId, address _nextId) external view override returns (address, address) {
+    function findInsertPosition(
+        uint256 _NICR,
+        address _prevId,
+        address _nextId
+    ) external view override returns (address, address) {
         return _findInsertPosition(troveManager, _NICR, _prevId, _nextId);
     }
 
-    function _findInsertPosition(ITroveManager _troveManager, uint256 _NICR, address _prevId, address _nextId) internal view returns (address, address) {
+    function _findInsertPosition(
+        ITroveManager _troveManager,
+        uint256 _NICR,
+        address _prevId,
+        address _nextId
+    ) internal view returns (address, address) {
         address prevId = _prevId;
         address nextId = _nextId;
 
@@ -386,11 +435,16 @@ contract SortedTroves is SortedTrovesStorage, CheckContract, ISortedTroves {
     // --- 'require' functions ---
 
     function _requireCallerIsTroveManager() internal view {
-        require(msg.sender == address(troveManager), "SortedTroves: Caller is not the TroveManager");
+        require(
+            msg.sender == address(troveManager),
+            "SortedTroves: Caller is not the TroveManager"
+        );
     }
 
     function _requireCallerIsBOorTroveM(ITroveManager _troveManager) internal view {
-        require(msg.sender == borrowerOperationsAddress || msg.sender == address(_troveManager),
-                "SortedTroves: Caller is neither BO nor TroveM");
+        require(
+            msg.sender == borrowerOperationsAddress || msg.sender == address(_troveManager),
+            "SortedTroves: Caller is neither BO nor TroveM"
+        );
     }
 }
